# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A URL shortener built with Go",
        "title": "Url shortener",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/api/v1",
    "paths": {
        "/analytics/{slug}": {
            "get": {
                "description": "Get total clicks and unique IP count for a URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Get URL analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug of the URL to get stats for",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analytics fetched successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "400": {
                        "description": "Slug is required",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Login a user with the provided credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginUserDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "description": "Logs out the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Logout user",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a new user with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.RegisterUserDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/delete/{slug}": {
            "delete": {
                "description": "Delete a shortened URL by its slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Delete a shortened URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug of the URL to delete",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "404": {
                        "description": "Slug not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/list": {
            "get": {
                "description": "List all shortened URLs for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "List user URLs",
                "responses": {
                    "200": {
                        "description": "List of user URLs",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "Create a shortened URL for the provided long URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "url"
                ],
                "summary": "Shorten a URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "url",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateURLDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Short URL created successfully",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        },
        "/{slug}": {
            "get": {
                "description": "Redirects the user to the original long URL for the given slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "redirect"
                ],
                "summary": "Redirect to long URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug of the shortened URL",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to the long URL"
                    },
                    "404": {
                        "description": "Slug not found",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dtos.StructuredResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateURLDto": {
            "description": "Url data for creating a short URL",
            "type": "object",
            "required": [
                "longurl"
            ],
            "properties": {
                "customalias": {
                    "description": "User's preffered alias\n@example abc123",
                    "type": "string",
                    "example": "abc123"
                },
                "longurl": {
                    "description": "User's original URL\n@example https://my-portfolio.com",
                    "type": "string",
                    "example": "https://my-portfolio.com"
                }
            }
        },
        "dtos.LoginUserDto": {
            "description": "Login credentials for authenticating a user",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "User's email address\n@example john.doe@example.com",
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "password": {
                    "description": "User's password\n@example SecureP@ssw0rd",
                    "type": "string",
                    "example": "SecureP@ssw0rd"
                }
            }
        },
        "dtos.RegisterUserDto": {
            "description": "Registration data for creating a new user account",
            "type": "object",
            "required": [
                "confirmPassword",
                "email",
                "password"
            ],
            "properties": {
                "confirmPassword": {
                    "description": "Confirmation of the password\n@example SecureP@ssw0rd",
                    "type": "string",
                    "example": "SecureP@ssw0rd"
                },
                "email": {
                    "description": "User's email address\n@example john.doe@example.com",
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "password": {
                    "description": "User's password (min 8 characters)\n@example SecureP@ssw0rd",
                    "type": "string",
                    "example": "SecureP@ssw0rd"
                }
            }
        },
        "dtos.StructuredResponse": {
            "description": "Standard response format containing success status, HTTP status code, message, and optional payload",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Human-readable message\n@example Operation completed successfully",
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "payload": {
                    "description": "Optional payload containing response data\n@example {\"id\":1,\"name\":\"Example Item\"}"
                },
                "status": {
                    "description": "HTTP status code\n@example 200",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "Whether the operation was successful\n@example true",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer: ` prefix, e.g. 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}